refactor(tests): split oversized test files at 400-line threshold
